tools/requirements.txt
requests
PyJWT
redis
